+++ Carbon/Callisto/Tests/Que.lua
@@ -1,0 +1,31 @@
+local Callisto = (...)
+local TestUtil = Callisto.TestUtil
+
+local settings = {}
+
+local source = [=[
+local obj = {
+	v = 5,
+	Que? = function(self)
+		return self.v
+	end,
+	?Por = function(self)
+		return self.v
+	end,
+	Help?Me = function(self)
+		return self.v
+	end
+}
+
+Output.a = obj:Que?()
+Output.b = obj:?Por()
+Output.c = obj:Help?Me()
+]=]
+
+local good_output = {
+	a = 5,
+	b = 5,
+	c = 5
+}
+
+return TestUtil.MakeTest(source, settings, good_output)