+++ Carbon/Testing/Tests/Carbide.lua
@@ -26,7 +26,7 @@ local function feature(test, name, ...)
 			error("Invalid feature definition!")
 		end
 
-		local source = Carbide.ParseCore(snippet, compilation_settings)
+		local source, err = Carbide.ParseCore(snippet, compilation_settings)
 		local compiled, err = Carbon.LoadString(source, "snippet #" .. i)
 
 		if (not compiled) then